Compute info hash on raw bencode before parsing
Previously we parsed the value of the info key, bound the result to a
struct, re-encoded the info struct and then created the hash. This
caused discrepencies.
Now we extract the raw bencode from the info and create a sha 1 from
that.

This commit introduces some tech debt that will have to be addressed:
Some bytes are being introduced after url encoding with ruby that we are
directly replacing now. This probably only works using the test torrent
file we have.
The url encoding also encodes '%' as '%25' which we are also replacing.