Get peers from bencoded tracker response
This commit parses the bencoded tracker response and returns a vector or
Peer structs. Peers have two attributes, ip and port.

The bencoded response includes a binary blob of the peers. The first 4
bytes are the array which we join into a string (We might be able to
represent this better, but will address when we are communicating with
out peers). The last 2 bytes is the port which is a big endian u16.

This commit caused immense grief as we could not parse the bencode. This
was due to the crate used for our http query handling. Reqwest would
return a string that mis-interpretted binary - it looks like it should
only really be used for plain text.
A binding of curl crate gave us the response as bytes which we could
pass into bencode::from_vec(CURL_RESPONSE).

We don't extract out the interval just yet as we don't need it. Interval
dictates how often we should query the tracker for an updated list of
peers. I can imagine we will have some thread of background job running
which will query every interval.