Make the .gitignore file ignore the 'target' directory and delete 'target/' from the worktree
The 'target' folder should never be committed into version control - the
files will change depending on which time the crate is compiled, and
which machine it is compiled on.

I think the .gitignore file was fine, you already fixed that, I just
removed a redundant line. Once you add something to the .gitignore, you
have to remove the file from the worktree to get it to be picked up.