Change test torrent to a more current torrent
When attempting to run the application a not very descriptive error
message explained that we could not unwrap none. This was happening when
tried to get a value from the bencoded response dict.

I changed the test torrent to a more current torrent and we could again
connection to peers and receive a bitfield.

TODO: Improve error message when we get an unsuccessful respoonse
FUTURE TODO: Retry initial handshake after x seconds